Add recombination map support in SLiM engine.
Closes #262.

Recombination map positions and rates are stored as arrays in the
SLiM script. As SLiM includes each script in the provenance table,
whole chromosome recombination maps result in very long arrays in
the provenance table. Note also that SLiM versions prior to 3.3.2
print the entire recombination map when initializeRecombinationRate()
is called.