Make mock unit tests less brittle, by using autospec.
For motivation, see passing tests that are obviously wrong:
https://github.com/popsim-consortium/stdpopsim/pull/475#discussion_r411214337

The problem is that mock.MagicMock() objects ensure all accessed
attributes and methods are successful. This means that unit tests
can do erroneous things, e.g. due to a typo or an API change, but the
test will pass regardless. A simple solution is to provide a `spec`
for the mocked object. In the case of mocking a function, this is
most easily done by passing `autospec=True` to the mock.patch()
or mock.patch.object() call. If a `new` parameter is provided for the
mock, the spec of the new object will be used, so these cases do not
suffer from the problem just described.

Note that autospec=True probably should not be applied to non-function
objects---at least not without first consulting the autospec docs.