Reduce readthedocs memory footprint.
Installing a conda environment on the readthdocs host often fails due
to memory exhaustion. But the readthedocs docker images don't include
gsl, which is needed to build msprime when installing with pip.
So this change uses a minimal conda environment, containing only gsl,
and installs the remaining dependencies using pip.

See #468.