Fix Zigzag_1S14 population sizes.
This makes the population sizes match those used in Schiffels & Durbin (2014),
and outputs a warning when simulating the Zigzag_1S14 model.

This also removes a test that fails on circleci. I was not able to
reproduce the failure offline, and github workflows tests also do not
reproduce the failure. The test adds no unique coverage.

Closes #745.